Create Chapter1
This chapter starts with a general introduction to microcontrollers and embedded systems and their definitions, as well as their importance and applications in our everyday lives. The chapter goes on to describe a simplified explanation of digital and analog electronics necessary for carrying out the microcontroller projects carried out in the rest of the chapters. This chapter also explains basic measurement equipment that may be used in this book’s projects. In this chapter you will quickly learn how to set up fundamental software tools for programming microcontrollers, as well as how to use basic electronic components as a starting point for programming examples for beginners.